% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdaview.R
\name{tdaview-shiny}
\alias{tdaview-shiny}
\alias{tdaviewOutput}
\alias{renderTdaview}
\title{Shiny bindings for tdaview}
\usage{
tdaviewOutput(outputId, width = "100\%", height = "400px")

renderTdaview(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit}

\item{expr}{An expression that generates a tdaview}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using tdaview within Shiny
applications and interactive Rmd documents.
}
