% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdaview.R
\name{tdaview}
\alias{tdaview}
\title{<TDAView>}
\usage{
tdaview(mapper, metadata, labels, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{mapper}{data frame containing adjacency matrix and vertices, formatted to match TDAMapper output}

\item{metadata}{single list of ordered categorical variables as strings}

\item{labels}{single list of ordered node labels as strings}
}
\description{
<Displays topological data as graph from TDAMapper or pHom>
}
