% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeMapper.R
\name{visualizeMapper}
\alias{visualizeMapper}
\title{Launch graph tool with Mapper input}
\usage{
visualizeMapper(mapper, metadata, labels = NULL)
}
\description{
Call this as an addin with required parameters to
}
