% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeMapper.R
\name{visualizeMapper}
\alias{visualizeMapper}
\title{Launch graph tool with Mapper input}
\usage{
visualizeMapper(mapper, metadata, names, labels = NULL)
}
\arguments{
\item{mapper}{data frame containing adjacency matrix and vertices, formatted to match TDAMapper output}

\item{metadata}{single list of ordered categorical variables as strings}

\item{names}{single list of ordered row names as strings from original data}

\item{labels}{single list of ordered node labels as strings}
}
\description{
Call this as an addin with required parameters to
}
